changed double type to decimal type

+++ holy-water/Bar.cs
@@ -5,13 +5,13 @@ namespace holy_water
     public class Bar
     {
         private String name;
-        private double volume, locX, locY;
+        private decimal volume, locX, locY;
         private int percentage;
         private int count;
-        private double average;
+        private decimal average;
 
         public int Count { get => count; }
-        public double Average { get => average; }
+        public decimal Average { get => average; }
 
         public int Percentage {
             get => percentage; 
@@ -22,11 +22,11 @@ namespace holy_water
         }
 
         public string Name { get => name; }
-        public double Volume { get => volume;  }
-        public double LocX { get => locX; }
-        public double LocY { get => locY; }
+        public decimal Volume { get => volume;  }
+        public decimal LocX { get => locX; }
+        public decimal LocY { get => locY; }
 
-        public Bar(string name, double volume, double locX, double locY, int percentage, double average = 0.0, int count = 1)
+        public Bar(string name, decimal volume, decimal locX, decimal locY, int percentage, decimal average = 0.0, int count = 1)
         {
             this.name = name;
             this.volume = volume;
